backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdfminer.six
opencv-python-headless
scikit-image
numpy
pyhanko
certifi
cryptography
pydantic
google-cloud-aiplatform
google-cloud-storage
python-dotenv

torch

transformers
accelerate
pillow

matplotlib
timm
yacs
scipy
setuptools
sqlalchemy
